+++ gears/src/types/base/coin/mod.rs
@@ -1,0 +1,5 @@
+mod decimal;
+mod unsigned;
+
+pub use decimal::*;
+pub use unsigned::*;
