+++ gears/src/types/base/coin/unsigned.rs
@@ -3,9 +3,7 @@ use serde::{Deserialize, Serialize};
 use std::str::FromStr;
 use tendermint::types::proto::Protobuf;
 
-use crate::types::{denom::Denom, errors::Error};
-
-use super::errors::CoinsError;
+use crate::types::{base::errors::CoinsError, denom::Denom, errors::Error};
 
 mod inner {
     pub use core_types::base::coin::Coin;
