+++ gears/src/types/base/mod.rs
@@ -1,6 +1,4 @@
 pub mod coin;
-pub mod decimal_coin;
-pub mod decimal_coins;
+pub mod coins;
 pub mod errors;
 pub mod min_gas;
-pub mod send;
