+++ gears/src/utils.rs
@@ -8,7 +8,7 @@ use crate::{
     },
     types::{
         address::AccAddress,
-        base::{coin::Coin, send::SendCoins},
+        base::{coin::Coin, coins::SendCoins},
         denom::Denom,
     },
 };
