+++ gears/src/x/ante.rs
@@ -5,7 +5,7 @@ use crate::signing::handler::MetadataGetter;
 use crate::signing::{handler::SignModeHandler, renderer::value_renderer::ValueRenderer};
 use crate::types::auth::gas::Gas;
 use crate::types::base::coin::Coin;
-use crate::types::base::send::SendCoins;
+use crate::types::base::coins::SendCoins;
 use crate::types::denom::Denom;
 use crate::types::gas::descriptor::{ANTE_SECKP251K1_DESCRIPTOR, TX_SIZE_DESCRIPTOR};
 use crate::types::gas::kind::TxKind;
